+++ src/main/java/cn/com/bgy/ifc/domain/impl/system/project/RegionStreetDomainImpl.java
@@ -1,14 +1,18 @@
 package cn.com.bgy.ifc.domain.impl.system.project;
 
+import cn.com.bgy.ifc.dao.system.project.RegionBuildingDao;
+import cn.com.bgy.ifc.dao.system.project.RegionComputerRoomDao;
 import cn.com.bgy.ifc.dao.system.project.RegionStreetDao;
 import cn.com.bgy.ifc.domain.interfaces.system.project.RegionStreetDomain;
 import cn.com.bgy.ifc.entity.po.system.project.RegionStreet;
 import cn.com.bgy.ifc.entity.vo.system.project.RegionStreetVo;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import javax.annotation.Resource;
 import java.util.Date;
 import java.util.List;
+import java.util.Map;
 
 /**
  * @Author huxin
@@ -20,6 +24,12 @@ public class RegionStreetDomainImpl implements RegionStreetDomain {
 
     @Resource
     private RegionStreetDao regionStreetDao;
+
+    @Resource
+    private RegionBuildingDao regionBuildingDao;
+
+    @Resource
+    private RegionComputerRoomDao regionComputerRoomDao;
     /**
      * @Author huxin
      * @Description 查
@@ -56,8 +66,26 @@ public class RegionStreetDomainImpl implements RegionStreetDomain {
      * @Description 删除
      * @Date 2018/12/19 15:32
      */
+    @Transactional
     @Override
     public int deleteRegionStreet( List<Long> list ) {
-        return regionStreetDao.deleteRegionStreet(list);
+        if(list.size()>0){
+            //删除机房
+            regionComputerRoomDao.deleteRegionComputerRoomBySuperId(list);
+            //删除楼栋
+            regionBuildingDao.deleteRegionBuildingBySuperId(list);
+            //删除区域
+            return regionStreetDao.deleteRegionStreet(list);
+        }
+        return 0;
+    }
+    /**
+     * @Author huxin
+     * @Description 根据父级id查询所有街道名
+     * @Date 2018/12/20 18:26
+     */
+    @Override
+    public List<Map<String, Object>> queryRegionStreetNameBySuperId( Long id ) {
+        return regionStreetDao.queryRegionStreetNameBySuperId(id);
     }
 }
